Update proton-pass extension
- Merge branch \'contributions/merge-1777542847754\' into codex/fix-proton-pass-vault-search
- Pull contributions
- Fix Proton Pass vault item normalization
- fix: keep hook order stable in search view
- Clarify independent proton-pass implementation and update changelog
- Implement pass-cli web login flow in Raycast auth UX
- Fix usePromise argument typing in search items
- Resolve changelog merge conflict
- Pull contributions
- Merge branch \'contributions/merge-1771858002740\'
- Pull contributions
- Add optional background refresh and web URL integration
- fix: restore Cmd+D shortcut, section grouping, and consistent debug copy (#1)
- fix: quarantine removal, default detail action, consistent copy toasts
- fix: resolve lint issue in totp refresh timer
- Merge branch \'contributions/merge-1771623325182\'
- Pull contributions
- chore: update proton pass command and metadata assets
- chore: update package-lock.json
- style: format pass-cli.ts with Prettier
- fix: address remaining Greptile review feedback
- chore: add CLAUDE.md to gitignore
- fix: address Greptile review feedback for security and correctness
- fix: add missing JSX import for error views component
- chore: update extension icon
- chore: update changelog, README, and add mock data for development
- docs: update README and clean up comments in code
- refactor: streamline error handling and improve code organization
- feat: add caching, vault filter, and improved error handling
- fix: align CLI integration with actual pass-cli JSON output
- Initial commit
